backend/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0

sqlalchemy
psycopg2-binary

python-jose[cryptography]
python-dotenv

requests

pydantic

apscheduler

python-dateutil

httpx